Agregar olvidar contraseña y cambios visuales a crear una cuenta
