+++ src/componentes/Firebase/inicar-sesion-contexto.js
@@ -1,10 +1,10 @@
 import React, { useRef } from "react";
 import { useContextoUsuario } from "../contexto/contextoUsuario";
-import { signInWithEmailAndPassword, sendPasswordResetEmail, } from "firebase/auth";
+import { signInWithEmailAndPassword } from "firebase/auth";
 import { auth } from "./firebase"
-import { Password } from 'primereact/password';
 import { InputText } from 'primereact/inputtext';
 import { Button } from 'primereact/button';
+
 
 const IniciarSesion = () => {
   const emailRef = useRef();
@@ -17,9 +17,7 @@ const IniciarSesion = () => {
     })
   }
 
-  const contraseñaOlvidada = (email) => {
-    return sendPasswordResetEmail(auth, email)
-  }
+
 
   const onSubmit = (e) => {
     e.preventDefault();
@@ -28,22 +26,15 @@ const IniciarSesion = () => {
     if (email && contraseña) iniciarSesion(email, contraseña);
   };
 
-  const forgotPasswordHandler = () => {
-    const email = emailRef.current.value;
-    if (email)
-      contraseñaOlvidada(email).then(() => {
-        emailRef.current.value = "";
-      });
-  };
+
 
   return (
     <div className="form">
-      <h1 className="text-blue-600"> Login </h1>
+      <h1 className="text-blue-600"> Iniciar Sesión </h1>
       <form onSubmit={onSubmit}>
         <InputText placeholder="Email" icon="pi pi-envelope" type="email" ref={emailRef} />
-        <Password placeholder="Contraseña" type="password" ref={contraseñaRef} />
-        <Button type="submit">Iniciar sesion</Button>
-        <p className="cursor-pointer" onClick={forgotPasswordHandler}>Contraseña olvidada?</p>
+        <InputText placeholder="Contraseña" type="password" ref={contraseñaRef} />
+        <Button type="submit">Iniciar Sesión</Button>
       </form>
     </div>
   );
