+++ src/componentes/bienvenida.js
@@ -7,7 +7,7 @@ import "primereact/resources/primereact.min.css"; //core css
 import "primeicons/primeicons.css";
 import SelectCategoria from "./categoria";
 import { useTranslation } from "react-i18next";
-import UserSidebar from "./firebase/user-sidebar";
+import UserSidebar from "./Firebase/user-sidebar";
 import CambiarIdioma from "./cambiar-idioma";
 import { useContextoUsuario } from "../componentes/contexto/contextoUsuario";
